FrontEnd-Prototype/requirements.txt
streamlit
tensorflow
pillow
numpy
opencv-python
rembg
scikit-image
matplotlib
onnxruntime
Pillow